python/backend/requirements.txt
absl_py==1.2.0
Flask==2.2.3
Flask_Cors==3.0.10
openai==0.27.2
gunicorn==20.1.0

